hibernate-orm-panache-kotlin in CI, test using injected resource extrated into separate class
